+++ src/itemAction/ItemAction.unit.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { mount, shallow } from 'enzyme'
 
 import { CrossIcon } from '../icon/crossIcon'
-import { ItemAction } from '../itemAction'
 import { Loader } from '../loader/Loader'
+import { ItemAction } from '.'
 
 jest.useFakeTimers()
